+++ tests/PeriodIteratorTest.php
@@ -3,6 +3,7 @@
 namespace iansltx\BusinessDays\Test;
 
 use iansltx\BusinessDays\FastForwarder;
+use iansltx\BusinessDays\Rewinder;
 use DateTimeImmutable as Immutable;
 use DateTime as Mutable;
 use DateInterval as Interval;
@@ -37,6 +38,36 @@ class PeriodIteratorTest extends \PHPUnit_Framework_TestCase
         $endDt = $ff->exec(new Mutable('2015-11-20 09:00:00'));
 
         $this->assertEquals('2015-12-07 09:00:00', $endDt->format('Y-m-d H:i:s'));
+        $this->assertInstanceOf('\DateTime', $endDt);
+    }
+
+    public function testRewindOverWeekend()
+    {
+        $rw = $this->addFilterSet1(Rewinder::createWithDays(2));
+
+        $endDt = $rw->exec(new Mutable('2015-02-10 09:00:00'));
+
+        $this->assertEquals('2015-02-06 09:00:00', $endDt->format('Y-m-d H:i:s'));
+        $this->assertInstanceOf('\DateTime', $endDt);
+    }
+
+    public function testRewindOverPresidentsDay()
+    {
+        $rw = $this->addFilterSet1(Rewinder::createWithDays(2));
+
+        $endDt = $rw->exec(new Mutable('2015-02-18 09:00:00'));
+
+        $this->assertEquals('2015-02-13 09:00:00', $endDt->format('Y-m-d H:i:s'));
+        $this->assertInstanceOf('\DateTime', $endDt);
+    }
+
+    public function testRewindOverBusinessDays()
+    {
+        $rw = $this->addFilterSet1(Rewinder::createWithDays(2));
+
+        $endDt = $rw->exec(new Mutable('2015-02-19 09:00:00'));
+
+        $this->assertEquals('2015-02-17 09:00:00', $endDt->format('Y-m-d H:i:s'));
         $this->assertInstanceOf('\DateTime', $endDt);
     }