+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All Notable changes to `iansltx\BusinessDays` will be documented in this file
 
 - Added bulk filter import/export to FastForwarder
 - Moved non-algorithmic functionality into SkipWhenTrait for use by future classes
+- Move algorithmic functionality into PeriodIterator class for use by future classes
 
 ## 1.2.0 (2015-02-16)