+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 All Notable changes to `iansltx\BusinessDays` will be documented in this file
 
-## NEXT
+## 1.3.0 (2015-02-18)
 
 - Added bulk filter import/export to FastForwarder
 - Added Rewinder for iterating backward from, rather than forward from, a given date
+- Added monthAndDaysOnDayOfWeek filter for, e.g., December 26 as long as it's a Friday
 - Moved non-algorithmic functionality into SkipWhenTrait for use by other potential classes
 - Fixed exec() return type when argument subclasses \DateTime or \DateTimeImmutable